requirements.txt
matplotlib>=3.5
pandas>=1.3
numpy>=1.21
jupyterlab
streamlit>=1.25
Pillow>=9.0
